+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Parse UserComment value when encoding is undefined but contains printable text.
+
 ## [4.38.1] - 2026-04-09
 
 ### Changed